pyproject.toml
[project]
name = "bill-splitter-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Ezequiel Saidman", email = "ezemati1904@gmail.com" }
]
requires-python = ">=3.14"
dependencies = [
    "argon2-cffi>=25.1.0",
    "fastapi[standard]>=0.136.0",
    "psycopg[binary,pool]>=3.3.3",
    "pydantic>=2.13.3",
    "pyjwt>=2.12.1",
    "sqlalchemy>=2.0.49",
]

[project.scripts]
bill-splitter-api = "bill_splitter_api:main"

[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[tool.fastapi]
entrypoint = "src.bill_splitter_api.main:app"
